chore: rework b2c login and organize wall components
