+++ src/js/wall/components/__tests__/FirstUser.test.tsx
@@ -1,9 +1,9 @@
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
+import { renderWithProviders } from "@tests/setupTests";
 import nock from "nock";
 import React from "react";
 import { describe, expect, it } from "vitest";
-import { renderWithProviders } from "../../../tests/setupTests";
 import FirstUser from "../FirstUser";
 
 describe("<FirstUser />", () => {
