+++ src/js/wall/components/__tests__/ResetForm.test.tsx
@@ -1,10 +1,10 @@
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
+import { renderWithProviders } from "@tests/setupTests";
 import nock from "nock";
 import React from "react";
 import { useDispatch } from "react-redux";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { renderWithProviders } from "../../../tests/setupTests";
 import ResetForm from "../ResetForm";
 
 vi.mock("react-redux", () => ({
